Make postinstall script
