+++ .gitignore
@@ -2,6 +2,9 @@
 /pc-nrfconnect-shared-*.tgz
 /coverage
 /scripts/nordic-publish.js
+/scripts/check-app-properties.js
+/scripts/check-for-typescript.js
+/scripts/nrfconnect-license.js
 /dist
 /typings/generated/*
 .vscode/settings.json
